requirements.txt
requests
python-dateutil
gspread
google-auth
PyYAML
openai